chore: commit all pending changes — data files, migrations, services, docs
Includes accumulated working changes across multiple sessions:
- Backend: API refinements (chat, energy, hvac, sites, sentry webhooks),
  service updates (AI optimizer, claude, embedding, hybrid AI, lifecycle,
  lighting sim, tier routing), circuit breaker, ML feeder, SceneCom test
- Frontend: Chat, Dashboard, SiteCard, KPICard, access control updates
- Data: audit log, consent records, parasite decisions, recommendations,
  sustainability metrics, niagara mappings/discoveries, capex analyses
- Migrations: renumbered 070-095 (resolved conflicts), new migrations
- Docs: Intent Engineering framework, REMS diagrams, API docs, testing
- Infra: SOPS encryption (.env.enc files), disk monitor, systemd units
- Phase 130 plan (SceneCom telemetry ingestion)

+++ supabase/migrations/046_water_data.sql
@@ -109,8 +109,8 @@ BEGIN
             ) VALUES (
                 'S002-MTR-W-MAIN',
                 'site-002',
-                ROUND(v_base_volume, 2),
-                ROUND(v_flow_rate, 2),
+                ROUND(v_base_volume::numeric, 2),
+                ROUND(v_flow_rate::numeric, 2),
                 v_date + (v_hour || ' hours')::INTERVAL,
                 FLOOR(v_base_volume / 10)::INTEGER, -- 10 liters per pulse
                 18 + (random() * 4), -- 18-22°C water temperature
